cast to float32 when loading content embs

+++ cdl/data.py
@@ -28,8 +28,8 @@ def load_content_embeddings(dataset_name, embedding=None, path=None, device=None
         raise Exception('Please specify either embedding model with --embedding,' 
         'or custom path of precomputed embeddings with --embedding_path')
     if path is None: path = f'data/processed/{dataset_name}/content-{embedding}.pt'
-    # Make sure dtype of loaded embeddings is double (float64)
-    x = torch.load(path, map_location=device).to(torch.double)
+    # Make sure dtype of loaded embeddings is float32, because it will be used by default for weight matrix
+    x = torch.load(path, map_location=device).to(torch.float32)
     if x.is_sparse:
         x = x.to_dense()
     return x
